model/requirements.txt
spacy
pillow
opencv-python
numpy
fuzzywuzzy
torch
matplotlib
scikit-image
pandas
torchvision